requirements.txt
# 基础依赖
python-dotenv==1.2.1
pydantic==2.12.5
PyYAML==6.0.3
numpy>=1.26.0

# 文件处理
PyPDF2==3.0.1
python-docx>=1.2.0
openpyxl>=3.1.5
pandas>=3.0.0
Pillow>=12.0.0
pytesseract==0.3.10

# UI框架
gradio>=6.0.0
gradio_client>=2.0.0

# 大语言模型
transformers==4.37.0
accelerate==1.12.0
bitsandbytes==0.49.1
cpm-kernels>=1.0.0
huggingface_hub>=0.36.0
tokenizers>=0.15.0

# Word文档生成
docxtpl==0.20.2

# 内存管理
psutil>=7.2.0

# PyTorch相关
torch>=2.3.0

# PyTorch扩展包
torchaudio>=2.3.0
torchvision>=0.18.0

# Web服务器相关
fastapi>=0.128.0
uvicorn>=0.40.0

# 其他依赖
filelock>=3.20.0
fsspec>=2025.12.0
scikit-learn==1.8.0
